refactor(git): reduce nodegit code surface area
