+++ src/git.ts
@@ -1,15 +1,7 @@
-import * as fs from 'fs';
-
-import * as git from 'nodegit';
-import * as moment from 'moment';
-
-import { Commit as GitCommit, Repository } from 'nodegit';
+import * as nodeGit from './git.nodegit';
 import { Commit } from './types';
 import logger from './logger';
-
-export function isShallowGitRepo(path: string): boolean {
-  return fs.existsSync(path + '.git/shallow');
-}
+import { isShallowGitRepo } from './git.utils';
 
 export async function getCommits({
   gitPaths,
@@ -38,129 +30,19 @@ export async function getCommits({
   );
 }
 
-export async function getCommitsForRepository({
-  gitPath,
-  countMerges,
-  since,
-  until,
-}: {
+export async function getCommitsForRepository(args: {
   gitPath: string;
   countMerges: boolean;
   since: string | Date;
   until: string | Date;
 }): Promise<Commit[]> {
-  if (isShallowGitRepo(gitPath)) {
-    logger.warn(`Cannot analyze shallow git repo: ${gitPath}!`);
+  if (isShallowGitRepo(args.gitPath)) {
+    logger.warn(`Cannot analyze shallow git repo: ${args.gitPath}!`);
     logger.warn(
-      `To fix this issue: run git fetch --unshallow inside ${gitPath}`,
+      `To fix this issue: run git fetch --unshallow inside ${args.gitPath}`,
     );
     return [];
   }
 
-  const repository: Repository = await git.Repository.open(gitPath);
-  const allReferences = await getAllReferences(repository);
-  const references = allReferences.filter((r) => r.match(/refs\/heads\/.*/));
-
-  const allCommits: Commit[] = [];
-  const latestBranchCommits: GitCommit[] = await Promise.all(
-    references.map((branchName) =>
-      getBranchLatestCommit(repository, branchName),
-    ),
-  );
-  const branchCommits = await Promise.all(
-    latestBranchCommits.map((latestCommit) =>
-      getBranchCommits({ latestCommit, since, until, repository }),
-    ),
-  );
-  branchCommits.forEach((commitList) =>
-    commitList.forEach((commit) => allCommits.push(commit)),
-  );
-
-  // Multiple branches might share commits, so take unique
-  const uniqueCommits = allCommits.filter(
-    generateUniquesBasedOnKeyFilter('sha'),
-  );
-
-  if (countMerges) {
-    return uniqueCommits;
-  }
-
-  const isNotMergeCommit = (commit) => !commit.message.startsWith('Merge ');
-  return uniqueCommits.filter(isNotMergeCommit);
+  return nodeGit.getCommitsForRepository(args);
 }
-
-function getAllReferences(repo: Repository): Promise<string[]> {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
-  return repo.getReferenceNames(git.Reference.TYPE.ALL);
-}
-
-export async function getBranchLatestCommit(
-  repo: Repository,
-  branchName: string,
-): Promise<GitCommit> {
-  const branch = await repo.getBranch(branchName);
-  return repo.getBranchCommit(branch.name());
-}
-
-export async function getBranchCommits({
-  latestCommit,
-  repository,
-  since,
-  until,
-}: {
-  latestCommit: GitCommit;
-  repository?: Repository;
-  since: string | Date;
-  until: string | Date;
-}): Promise<Commit[]> {
-  if (!repository) {
-    // eslint-disable-next-line no-param-reassign
-    repository = latestCommit.owner();
-  }
-  return new Promise((resolve, reject) => {
-    const history = latestCommit.history();
-    const commits = [];
-
-    history.on('commit', (commit) => {
-      const commitAuthor = commit.author();
-      const author = commitAuthor
-        ? { name: commitAuthor.name(), email: commitAuthor.email() }
-        : { name: 'unknown', email: 'unknown@example.com' };
-
-      const commitData: Commit = {
-        sha: commit.sha(),
-        date: commit.date(),
-        message: commit.message(),
-        author,
-        repo: repository.commondir().replace('/.git/', ''),
-      };
-
-      const commitDate = moment(commitData.date.toISOString());
-      const sinceAlways = since === 'always' || !since;
-      const isValidSince = sinceAlways || commitDate.isAfter(since);
-
-      const untilAlways = until === 'always' || !until;
-      const isValidUntil = untilAlways || commitDate.isBefore(until);
-
-      if (isValidSince && isValidUntil) {
-        commits.push(commitData);
-      }
-    });
-
-    history.on('end', () => resolve(commits));
-    history.on('error', (err) => reject(err));
-    history.start();
-  });
-}
-
-const generateUniquesBasedOnKeyFilter = (key: string) => {
-  const uniques = new Set();
-  return (item) => {
-    const alreadyAdded = uniques.has(item[key]);
-    if (!alreadyAdded) {
-      uniques.add(item[key]);
-    }
-    return !alreadyAdded;
-  };
-};
